requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
sqlmodel==0.0.14
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
psycopg2-binary==2.9.10
pydantic-settings==2.9.1
python-multipart==0.0.20